svcs_mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "svcs-mcp"
version = "1.0.0"
description = "SVCS MCP Server - Semantic Version Control System with Model Context Protocol support"
readme = "README.md"
authors = [
    {name = "SVCS Contributors", email = "svcs@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["git", "semantic-analysis", "code-evolution", "mcp", "ai"]
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/svcs/svcs-mcp"
Documentation = "https://svcs-mcp.readthedocs.io"
Repository = "https://github.com/svcs/svcs-mcp.git"
Issues = "https://github.com/svcs/svcs-mcp/issues"

[project.scripts]
svcs = "svcs_mcp.cli:main"
svcs-mcp-server = "svcs_mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["svcs_mcp*"]

[tool.setuptools.package-data]
svcs_mcp = ["templates/*", "hooks/*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
